medical-rag-ai/frontend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
langchain-google-genai
langchain-core
pydantic
pypdf
Pillow
python-dotenv
streamlit
requests